refactor: modify add provider layout and integrate Ad folder into Announcement folder
