+++ kagami/webui/src/page/AdminDashboard/WorkerDetail/styles.module.scss
@@ -122,13 +122,26 @@
   border-radius: 8px;
   padding: 20px;
 
-  h3 {
-    color: #fff;
-    margin: 0 0 16px 0;
-    font-size: 18px;
-    font-weight: 500;
+  .resourceHeader {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 16px;
     padding-bottom: 12px;
     border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+
+    h3 {
+      color: #fff;
+      margin: 0;
+      font-size: 18px;
+      font-weight: 500;
+    }
+
+    .resourceHeaderActions {
+      display: flex;
+      align-items: center;
+      gap: 12px;
+    }
   }
 }
 
@@ -317,7 +330,7 @@
     color: #999;
   }
 
-  input {
+  input, select {
     width: 100%;
     padding: 8px;
     border: 1px solid rgba(255, 255, 255, 0.1);
@@ -328,6 +341,14 @@
     &:focus {
       outline: none;
       border-color: #2196f3;
+    }
+  }
+
+  select {
+    cursor: pointer;
+    option {
+      background: #2a2a2a;
+      color: #fff;
     }
   }
 }
@@ -344,7 +365,7 @@
     font-size: 14px;
     font-weight: 500;
     cursor: pointer;
-    transition: all 0.2s ease;
+    transition: all 0.2s;
     border: none;
     outline: none;
