+++ kagami/webui/vite.config.ts
@@ -7,6 +7,6 @@ export default defineConfig(() => ({
   },
   plugins: [react()],
   server: {
-    port: 3000,
+    port: 11451,
   },
 }));
